requirements.txt
asgiref==3.4.1
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.1.0
Babel==2.9.1
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.9.3
billiard==4.2.0
blinker==1.4
boto3==1.10.41
botocore==1.13.50
cachetools==5.3.3
cairocffi==1.7.0
CairoSVG==2.7.1
cffi==1.15.0
chardet==3.0.4
click==8.0.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
cron-descriptor==1.4.3
cryptography==2.8
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.4
django-3-jet==1.0.7
django-celery-beat==2.6.0
django-cors-headers==3.7.0
django-crontab==0.7.1
django-db-logger==0.1.13
django-extensions==2.2.3
django-filter==2.4.0
django-google-maps==0.12.4
django-import-export==2.5.0
django-ipware==4.0.2
django-localflavor==2.2
django-mysql==3.12.0
django-phonenumber-field==3.0.1
django-rest-auth==0.9.5
django-silk==5.1.0
django-simple-history==3.0.0
django-storages==1.8
django-taggit==1.3.0
django-timezone-field==6.1.0
django-user-agents==0.4.0
djangorestframework==3.12.4
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.20.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.1
Flask==2.0.1
future==0.18.2
geographiclib==2.0
geopy==2.4.0
gprof2dot==2022.7.29
gspread==5.9.0
gspread-dataframe==3.3.1
h11==0.14.0
html5lib==1.1
httplib2==0.15.0
idna==2.8
iniconfig==2.0.0
ipython==7.16.3
itsdangerous==2.0.1
itypes==1.2.0
jedi==0.17.2
jeepney==0.8.0
Jinja2==3.0.3
jmespath==0.10.0
keyring==18.0.1
kombu==5.3.7
lxml==4.7.1
MarkupPy==1.14
MarkupSafe==2.0.1
msgpack==1.0.8
num2words==0.5.12
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.46
openpyxl==3.0.9
outcome==1.3.0.post0
packaging==24.0
pandas==1.1.5
paramiko==2.6.0
parso==0.7.1
pdfrw==0.4
pexpect==4.9.0
phonenumbers==8.10.20
pickleshare==0.7.5
pika==1.3.1
Pillow==8.2.0
pluggy==1.5.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.10.0
pyhtml2pdf==0.0.6
PyJWT==2.8.0
pymongo==3.12.1
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.6
pypdf==3.8.1
pyphen==0.15.0
pytest==6.2.5
python-crontab==3.0.0
python-dateutil==2.7.3
python-dotenv==1.0.1
python-stdnum==1.17
pytz==2019.3
PyYAML==5.3.1
razorpay==1.3.0
redis==5.0.0
requests==2.23.0
requests-oauthlib==2.0.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
Rx==3.2.0
s3transfer==0.2.1
SecretStorage==3.3.3
simplejson==3.16.0
six==1.14.0
slugify==0.0.1

sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
uritemplate==4.1.1
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
WeasyPrint==52.5
webdriver-manager==4.0.1
webencodings==0.5.1
Werkzeug==2.0.2
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
